requirements.txt
git+https://github.com/hsum/pybaseball.git@fix-pandas-concat-warning#egg=pybaseball
polars>=0.20.0
fastapi>=0.112.0
uvicorn>=0.30.0
pydantic>=2.8.0
pytest>=8.3.0
ruff>=0.6.0
boto3>=1.34.0
click>=8.1.0
httpx>=0.27.0
python-dotenv>=1.0.0
